Tests
Placed at 0x7000. From the project documentation, I copied over all
op-codes we are required to be able to disassemble. Then, starting from
MOVE and working my way down, I wrote test data with various input for
<source> and <destination> to cover the effective addressing modes we
are required to disassemble. I also applied this to the size ranges each
instruction can address. I'm not finished and the last of the op-code
instructions are commented out. In this way I hope to address most
possible combinations. After we get the basics, it would be good to
throw in a wrench or two.