Finished test op-codes
Coded all possible op-code and effective address combinations that will
be tested, including varying sizes, with the exception of varying the
register numbers, direct data input, or direct addresses. The end
product is a large section of code which aims to break the disassembler
and exploit any possible errors in logic.