Successful decode of NOP
Established bit shift and isolation methods, still needs to be
modularized, but is currently effective. May need some tweaking due to
unforeseeable requirements with other op-codes. Current setup isolates
bits to be tested, then branches and tests more bits until paired down
to a single op-code that can be pushed to buffer. Each branching scheme
ends with a BNE (not equal) where it will eventually push the generic
'DATA' term to the buffer.