Prints out most all op-codes now
It still has a hang at the VERY VERY end. It was hanging because when it
encountered a word it didn't know, the op-code routine was branching to
append generic data. But since it was a branch and not a subroutine,
when it hit RTS at the end of the branch, the program would halt. I
commented out some of your code and added some other lines. I just
wasn't sure what the code in EA was doing and I wanted to understand why
it was printing out so much extra stuff. You can uncomment or delete my
added code if need be. Just didn't want to delete your code without know
it's purpose. Also, I changed your prtBuf routine to overwrite the
information in the buffer right after printing.

The new OCR test is also much more comprehensive, covering all the
op-codes. Any questions, please ask. I have a really good understanding
of what is going on now. Long night. Off to bed.