Prints effective addresses correctly, but it doesn't print all of them.
Lines 183 and 196: Added save/restore register state to prtBuf.

Line 186: Changed the trap task that executes in prtBuf. Trap task goes
to address in buffer, goes to string that address points to, then
outputs everything up to the NULL termination of the string. Prints more
than we want.

Line 447: Changed the design of ADDR_REGS and DATA_REGS. With the
previous design, the program was accessing Words, but values were bytes,
so it would print two of them because they were stored in memory back to
back.

Line 468: I increased the ascii buffer size to an arbitrary number
because the value was too small.

Line 286: Pushes a NULL character to the ascii buffer since Trap task
#14 needs it, or it prints more than what is in the buffer.